+++ CHANGELOG.md
@@ -4,13 +4,14 @@
 
 ### BC breaks
 
-* changed `stubbles\webapp\interceptor\PreInterceptor::preProcess()` and `stubbles\webapp\interceptor\PostInterceptor::postProcess()` to always return a boolean value
+* Changed `stubbles\webapp\interceptor\PreInterceptor::preProcess()` and `stubbles\webapp\interceptor\PostInterceptor::postProcess()` to always return a boolean value
 
 ### Other changes
 
 * Added `stubbles\webapp\request\WebRequest::uploads()` which provides access to uploaded files
 * Added `image/jpeg` as default supported mime type when stubbles/image is present
 * `stubbles\webapp\auth\InternalAuthProviderException` and `stubbles\webapp\auth\ExternalAuthProviderException` now accept `\Throwable` as cause instead of `\Exception` only
+* Fixed type error when user agent name string is `null`
 
 ## 8.0.1 (2019-11-21)